chore(project): renamed namespaces
