+++ OnlineGameStore.DAL/Configurations/GameConfiguration.cs
@@ -1,5 +1,6 @@
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
+using OnlineGameStore.DAL.Entities;
 
 namespace OnlineGameStore.DAL;
