+++ OnlineGameStore.DAL/Interfaces/IGameRepository.cs
@@ -1,3 +1,5 @@
-﻿namespace OnlineGameStore.DAL;
+﻿using OnlineGameStore.DAL.Entities;
+
+namespace OnlineGameStore.DAL.Interfaces;
 
 public interface IGameRepository : IRepository<Game> { }
